Package restructure reviewed
