--- src/main/java/com/thiccWallet/api/common/exception/DuplicateCredentialsException.java
@@ -1,7 +1,0 @@
-package com.thiccWallet.api.common.exception;
-
-public class DuplicateCredentialsException extends RuntimeException {
-    public DuplicateCredentialsException(String msg) {
-        super(msg);
-    }
-}
