--- src/main/java/com/thiccWallet/api/common/exception/DuplicateLoginAttemptException.java
@@ -1,7 +1,0 @@
-package com.thiccWallet.api.common.exception;
-
-public class DuplicateLoginAttemptException extends RuntimeException {
-    public DuplicateLoginAttemptException(String msg) {
-        super(msg);
-    }
-}
