--- src/main/java/com/thiccWallet/api/common/exception/InvalidRequestException.java
@@ -1,7 +1,0 @@
-package com.thiccWallet.api.common.exception;
-
-public class InvalidRequestException extends RuntimeException {
-    public InvalidRequestException(String msg) {
-        super(msg);
-    }
-}
