--- src/main/java/com/thiccWallet/api/common/exception/NoSuchUserException.java
@@ -1,7 +1,0 @@
-package com.thiccWallet.api.common.exception;
-
-public class NoSuchUserException extends RuntimeException {
-    public NoSuchUserException(String msg) {
-        super(msg);
-    }
-}
