--- src/main/java/com/thiccWallet/api/common/exception/NotLoggedInException.java
@@ -1,7 +1,0 @@
-package com.thiccWallet.api.common.exception;
-
-public class NotLoggedInException extends RuntimeException {
-    public NotLoggedInException(String msg) {
-        super(msg);
-    }
-}
