+++ src/main/java/com/thiccWallet/api/data/user/dtos/responses/UserResponse.java
@@ -1,0 +1,66 @@
+package com.thiccWallet.api.data.user.dtos.responses;
+
+import com.thiccWallet.api.data.user.User;
+
+import java.time.LocalDateTime;
+
+public class UserResponse {
+
+    private String userId;
+    private String username;
+    private String email;
+    private LocalDateTime dateCreated;
+
+    public UserResponse(User user) {
+        this.userId = user.getId();
+        this.username = user.getUsername();
+        this.email = user.getEmail();
+        this.dateCreated = user.getDateCreated();
+    }
+
+    public UserResponse() {
+        super();
+    }
+
+    public String getUserId() {
+        return userId;
+    }
+
+    public void setUserId(String userId) {
+        this.userId = userId;
+    }
+
+    public String getUsername() {
+        return username;
+    }
+
+    public void setUsername(String username) {
+        this.username = username;
+    }
+
+    public String getEmail() {
+        return email;
+    }
+
+    public void setEmail(String email) {
+        this.email = email;
+    }
+
+    public LocalDateTime getDateCreated() {
+        return dateCreated;
+    }
+
+    public void setDateCreated(LocalDateTime dateCreated) {
+        this.dateCreated = dateCreated;
+    }
+
+    @Override
+    public String toString() {
+        return "UserResponse{" +
+                "userId='" + userId + '\'' +
+                ", username='" + username + '\'' +
+                ", email='" + email + '\'' +
+                ", dateCreated=" + dateCreated +
+                '}';
+    }
+}
