--- src/main/java/com/thiccWallet/api/login/dtos/request/LoginRequest.java
@@ -1,44 +1,0 @@
-package com.thiccWallet.api.login.dtos.request;
-
-import javax.validation.constraints.NotBlank;
-
-public class LoginRequest {
-
-    @NotBlank
-    private String username;
-    @NotBlank
-    private String password;
-
-    public LoginRequest() {
-
-    }
-
-    public LoginRequest(String username, String password) {
-        this.username = username;
-        this.password = password;
-    }
-
-    public String getUsername() {
-        return username;
-    }
-
-    public void setUsername(String username) {
-        this.username = username;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
-
-    @Override
-    public String toString() {
-        return "LoginRequest{" +
-                "username='" + username + '\'' +
-                ", password='" + password + '\'' +
-                '}';
-    }
-}
