--- src/main/java/com/thiccWallet/api/models/user/UserRepository.java
@@ -1,18 +1,0 @@
-package com.thiccWallet.api.models.user;
-
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.CrudRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-@Repository
-public interface UserRepository extends CrudRepository<User, String> {
-
-    Optional<User> findUserByUsername(String username);
-    Optional<User> findUserByEmail(String email);
-
-    @Query("from User u where u.username = :username and u.password = :password")
-    Optional<User> findUserByUsernameAndPassword(String username, String password);
-
-}
