--- src/main/java/com/thiccWallet/api/models/user/dtos/requests/UserEditRequest.java
@@ -1,51 +1,0 @@
-package com.thiccWallet.api.models.user.dtos.requests;
-
-public class UserEditRequest {
-
-    private String username;
-    private String password;
-    private String email;
-
-    public UserEditRequest() {
-
-    }
-
-    public UserEditRequest(String username, String password, String email) {
-        this.username = username;
-        this.password = password;
-        this.email = email;
-    }
-
-    public String getUsername() {
-        return username;
-    }
-
-    public void setUsername(String username) {
-        this.username = username;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    @Override
-    public String toString() {
-        return "UserEditRequest{" +
-                "username='" + username + '\'' +
-                ", password='" + password + '\'' +
-                ", email='" + email + '\'' +
-                '}';
-    }
-}
