+++ src/main/java/com/thiccWallet/api/util/common/exception/DuplicateCredentialsException.java
@@ -1,0 +1,7 @@
+package com.thiccWallet.api.util.common.exception;
+
+public class DuplicateCredentialsException extends RuntimeException {
+    public DuplicateCredentialsException(String msg) {
+        super(msg);
+    }
+}
