+++ src/main/java/com/thiccWallet/api/util/common/exception/DuplicateLoginAttemptException.java
@@ -1,0 +1,7 @@
+package com.thiccWallet.api.util.common.exception;
+
+public class DuplicateLoginAttemptException extends RuntimeException {
+    public DuplicateLoginAttemptException(String msg) {
+        super(msg);
+    }
+}
