+++ src/main/java/com/thiccWallet/api/util/common/exception/InvalidRequestException.java
@@ -1,0 +1,7 @@
+package com.thiccWallet.api.util.common.exception;
+
+public class InvalidRequestException extends RuntimeException {
+    public InvalidRequestException(String msg) {
+        super(msg);
+    }
+}
