+++ src/main/java/com/thiccWallet/api/util/common/exception/NoSuchUserException.java
@@ -1,0 +1,7 @@
+package com.thiccWallet.api.util.common.exception;
+
+public class NoSuchUserException extends RuntimeException {
+    public NoSuchUserException(String msg) {
+        super(msg);
+    }
+}
