+++ src/main/java/com/thiccWallet/api/util/common/exception/NotLoggedInException.java
@@ -1,0 +1,7 @@
+package com.thiccWallet.api.util.common.exception;
+
+public class NotLoggedInException extends RuntimeException {
+    public NotLoggedInException(String msg) {
+        super(msg);
+    }
+}
